Stage/backend/requirements.txt
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
